Rename wasser to backport_request
Also, autopep8 tests.
